+++ CHANGES.rst
@@ -1,3 +1,11 @@
+1.1.2
+=====
+* [ENHANCEMENT] Client can now choose alternative when async is not ideal
+
+1.1.1
+=====
+* [ENHANCEMENT] Redis Sentinal Support
+
 1.1.0
 =====
 * [ENHANCEMENT] Add this file